revendedora

+++ src/main/java/br/com/claricejoias_ws/dto/LeadItemDTO.java
@@ -5,12 +5,14 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class LeadItemDTO {
 
     private Long id;
     private Integer quantidade;
-    private Double precoMomento;
+    private BigDecimal precoMomento;
     private ProdutoDTO produto;
 }
